mcp_usecases/17_mcp_gmail/pyproject.toml
[project]
name = "17-mcp-gmail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.2.4",
    "google-api-python-client>=2.194.0",
    "google-auth>=2.49.2",
    "google-auth-oauthlib>=1.3.1",
    "mcp[cli]>=1.27.0",
    "python-dotenv>=1.2.2",
]
